third_party/WebKit/LayoutTests/fast/parser/entity-end-script-tag.html

<!DOCTYPE html>

<script src="../../resources/js-test.js"></script>
<p id="description"></p>
<div id="console"></div>
<script>
description("Test parsing of entity-escaped &lt;/script&gt; tag for <a href=\"https://bugs.webkit.org/show_bug.cgi?id=7931\">Bug 7931: Escaped elements within a textarea block can cause the textarea box to be closed prematurely</a>");

expectError();
</script>
<script id="myscript">&lt;/script></script>
<script>
shouldHaveHadError("SyntaxError");

var expectedResult = '&lt;' + '/script>';
shouldBe('document.getElementById("myscript").textContent', 'expectedResult');
</script>
